requirements.txt
numpy
scipy
pandas
sklearn
matplotlib
tqdm
rich
pytorch-lightning
hydra-core
omegaconf
wandb
einops
opt_einsum
cmake # For pykeops support
pykeops # If there are installation problems with pykeops==2.x, try pykeops==1.5
transformers # For some schedulers

# Model specific packges
# pytorch-fast-transformers # for Performer

# Dataset specific packages
torchtext # LRA
datasets # LRA
# gluonts
# timm
# lightning-bolts
# sktime # BIDMC
